# docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Pixel WEB SERVER",
    "version": "1.0.0",
    "description": "Документація до беку"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "https",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/products": {
      "get": {
        "tags": [
          "Anon"
        ],
        "summary": "Вивід продуктів на клієнті",
        "description": "Цей запит дозволяє отримати список продуктів з підтримкою фільтрів, пагінації та можливістю пошуку за ID продукту",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "ID продукту для отримання одного продукту",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Номер сторінки для пагінації",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "price",
            "in": "query",
            "description": "Фільтр за ціною",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "rate",
            "in": "query",
            "description": "Фільтр за рейтингом",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "promotion",
            "in": "query",
            "description": "Фільтр за наявністю акції",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sectionId",
            "in": "query",
            "description": "ID секції для фільтрації за секцією",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "subSectionId",
            "in": "query",
            "description": "ID підсекції для фільтрації за підсекцією",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Успішна відповідь із продуктами",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "products": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "name": {
                            "type": "string"
                          },
                          "price": {
                            "type": "number"
                          },
                          "rate": {
                            "type": "number"
                          },
                          "promotion": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "currentPage": {
                      "type": "integer",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "example": 5
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Продукт не знайдено",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Not Found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Помилка сервера",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}